refactor:アプリケーション構造の整理 - 録音機能の古いverを分離 - app直下にログインページを異動
